- create Makefile - Restructure header files: move all logic to .c file, leave .h files to be REAL header file
To compile: make
