+++ .gitignore
@@ -2,4 +2,4 @@
 *.out
 *.o
 *.a
-assignment*
+chat_app
